app: extract the application entrypoint
